+++ opsgenie/resource_opsgenie_integration_action_test.go
@@ -207,8 +207,12 @@ resource "opsgenie_integration_action" "test_api" {
 	message = "{{message}}"
 	description = "{{description}}"
 	entity = "{{entity}}"
+    priority = "P2"
 	alert_actions = ["Check error rate"]
-    
+    extra_properties = map(
+      "Environment", "test-env",
+      "Region", "us-west-2"
+    )
     filter {
       type = "match-all-conditions"
       conditions {
@@ -224,6 +228,7 @@ resource "opsgenie_integration_action" "test_api" {
   }
   create {
     name = "Create medium priority alerts"
+	order = 2
     tags = ["SEVERE", "SEV-1"]
     filter {
       type = "match-all-conditions"