+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.4.3 (August 14, 2020)
+
+IMPROVEMENTS:
+* **Integration Actions :** add priority for create action (#157)
+
+## 0.4.2 (August 13, 2020)
+
+BUGFIX:
+* **Integration Actions :** allow extra_properties (#152)
+
+## 0.4.1 (July 29, 2020)
+
+Opsgenie Provider repository changes
+
 ## 0.4.0 (July 28, 2020)
 
 IMPROVEMENTS: